Add navigation

+++ feature/characters/src/main/java/io/fajarca/characters/domain/MarvelCharacter.kt
@@ -1,3 +1,3 @@
 package io.fajarca.characters.domain
 
-data class MarvelCharacter(val id: Long, val title: String, val imageUrl: String)
+data class MarvelCharacter(val id: Int, val title: String, val imageUrl: String)
